myhackernews/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
babel==2.16.0
backoff==2.2.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
courlan==1.3.0
dateparser==1.2.0
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
haystack-ai==2.4.0
haystack-experimental==0.1.1
htmldate==1.8.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jusText==3.0.1
lazy-imports==0.3.1
lxml==5.3.0
lxml_html_clean==0.2.0
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.4.0
networkx==3.3
numpy==1.26.4
openai==1.42.0
pandas==2.2.2
posthog==3.5.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tenacity==9.0.0
tld==0.13
tqdm==4.66.5
trafilatura==1.12.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.4
